[feat]: Windows build pipeline, one version source and host telemetry
Windows x86_64 builds and runs all three editions (slim, full+cuda,
full+vulkan) through scripts\setup.ps1, native-libs\build-all.ps1 and
build.ps1: MSVC located with vswhere, Ninja, Git Bash for the shared shell
scripts, an isolated whisper_tf.dll linked from a generated .def, the
prebuilt zvec SDK, and /NODEFAULTLIB:libcmt.lib where CUDA's cuda.lib
requests LIBCMT. The executing side of tentanas-helper and the teams-bot
signal handling are cfg(unix), so the catalog still links everywhere.

Every native library ref, checksum, toolkit and SDK version now lives in
scripts/versions.env, read by scripts/lib/versions.sh (bash),
scripts/lib/windows.ps1 (PowerShell), addons/Directory.Build.props (MSBuild)
and tentaflow-core/build.rs. Pins are capped by their consumers: CUDA 13.3
(cudarc), WASI SDK 29.0 (NativeAOT-LLVM), TensorRT 10 (the ORT EP links
nvinfer_10). sherpa-rs is replaced by our own VITS wrapper over
sherpa-rs-sys.

Host telemetry: gpu_telemetry samples DXGI adapters and PDH usage once a
second and feeds node metrics, the service VRAM hint and the profiler.
nvidia-smi reports [N/A] per process under WDDM, so such rows are dropped
instead of counted as zero and Windows reads per-process VRAM from OS
accounting. Cards are matched to a tool's row by name and ordinal, never by
index. Every PDH counter goes through PdhAddEnglishCounterW with wildcard
instances -- the localized names matched nothing on a non-English Windows,
so the CPU, disk, RAM and GPU collectors returned no samples at all.
Interface link state, MAC, medium, speed and gateway come from netdev off
Linux. Node metrics also refresh without a mesh, the firewall self-check no
longer holds up startup on its UAC prompt, and ML Studio GPU stats read the
shared collector instead of spawning nvidia-smi.

+++ tentaflow-wrappers/src/llama.rs
@@ -60,8 +60,10 @@ const LOG_THRESHOLD_SILENT: u32 = u32::MAX;
 
 // Globalny prog poziomu logow llama.cpp/ggml. Inicjalizowany RAZ z env
 // `TENTAFLOW_LLAMA_LOG_LEVEL`; `LOG_THRESHOLD_SILENT` = pelne wyciszenie.
+// Trzymany jako u32: bindgen daje enum `ggml_log_level` jako u32 na Linuksie i
+// i32 na MSVC, a jego wartosci (0..=5) mieszcza sie w obu.
 #[cfg(feature = "llama")]
-static LOG_THRESHOLD: AtomicU32 = AtomicU32::new(sys::GGML_LOG_LEVEL_WARN);
+static LOG_THRESHOLD: AtomicU32 = AtomicU32::new(sys::GGML_LOG_LEVEL_WARN as u32);
 
 #[cfg(feature = "llama")]
 static LOG_THRESHOLD_INIT: Once = Once::new();
@@ -75,13 +77,13 @@ fn ensure_log_threshold() {
         let threshold = match std::env::var("TENTAFLOW_LLAMA_LOG_LEVEL") {
             Ok(value) => match value.trim().to_ascii_lowercase().as_str() {
                 "none" => LOG_THRESHOLD_SILENT,
-                "error" => sys::GGML_LOG_LEVEL_ERROR,
-                "warn" => sys::GGML_LOG_LEVEL_WARN,
-                "info" => sys::GGML_LOG_LEVEL_INFO,
-                "debug" => sys::GGML_LOG_LEVEL_DEBUG,
-                _ => sys::GGML_LOG_LEVEL_WARN,
+                "error" => sys::GGML_LOG_LEVEL_ERROR as u32,
+                "warn" => sys::GGML_LOG_LEVEL_WARN as u32,
+                "info" => sys::GGML_LOG_LEVEL_INFO as u32,
+                "debug" => sys::GGML_LOG_LEVEL_DEBUG as u32,
+                _ => sys::GGML_LOG_LEVEL_WARN as u32,
             },
-            Err(_) => sys::GGML_LOG_LEVEL_WARN,
+            Err(_) => sys::GGML_LOG_LEVEL_WARN as u32,
         };
         LOG_THRESHOLD.store(threshold, Ordering::Relaxed);
     });
@@ -98,7 +100,7 @@ unsafe extern "C" fn filtered_log(
     _user_data: *mut std::os::raw::c_void,
 ) {
     let threshold = LOG_THRESHOLD.load(Ordering::Relaxed);
-    if threshold == LOG_THRESHOLD_SILENT || level < threshold {
+    if threshold == LOG_THRESHOLD_SILENT || (level as u32) < threshold {
         return;
     }
     if text.is_null() {
